Refactor: 위치변경, 최종 수정
